Deal with broken feeds
The cache.set tries to pickle the feed object.  If that contains a
bozo_exception element (which will be an exception) then the cache will
throw an error and the entire page will be a 500.

So now we check for bozo_exception and if it's there we log it and
delete it.