made it possible to set the title to the empty string
the help text of this field already pretended that it is possible to leave the explicit title out (such that the feed's title is used instead). But null=True is not enough, it also needs blank=True. This commits fixes this issue.

Beside that, the verbose_name and help_text attributes are slightly improved.