moved the template choices from the model to the plugin form
this is important because making the CharField choices contraint in the model dependent on (user) settings means that the migration system wants to create a new migration each time the settings variable changes. Such a migration would include the hardcoded names of the user's templates which is definitely unintended behavior.